Revisi SKU Product as Integer

+++ applications/dao/AuditDao.py
@@ -121,7 +121,7 @@ def getDataAuditByFaktur(faktur):
             price,
             tt.qty * price as subtotal
         FROM tx_trans_detail_audit tt
-            INNER JOIN ms_product mp on tt.sku = mp.sku
+            INNER JOIN ms_product mp on tt.sku::integer = mp.sku::integer
             INNER JOIN ms_category mc on mc.category_id = mp.category_id
         WHERE faktur = %(faktur)s
         ORDER BY faktur;
